Refactore and improve stricutre code
