+++ src/field.js
@@ -1,7 +1,7 @@
-import { getErrorMessage, contextWrapper } from "./utils";
-import * as rules from "./rules";
+import { getErrorMessage, fieldWrapper } from "./utils";
+import rules from "./rules/index";
 
-class Context {
+class Field {
   #rules = [];
 
   _add(rule) {
@@ -28,4 +28,4 @@ class Context {
   }
 }
 
-export default contextWrapper(rules)(Context);
+export default fieldWrapper(rules)(Field);
