+++ src/index.js
@@ -1,6 +1,6 @@
-import Context from "./context";
+import Field from "./field";
 import Schema from "./schema";
 
-export const field = () => new Context();
+export const field = () => new Field();
 
 export const typeSchema = Schema;
