+++ src/rules/number.js
@@ -1,0 +1,9 @@
+export const number = (options = {}) => ({
+  validator: value => typeof value === "number",
+  options
+});
+
+export const between = (minValue, maxValue, options = {}) => ({
+  validator: value => minValue <= value && value <= maxValue,
+  options
+});
