+++ src/rules/object.js
@@ -1,0 +1,6 @@
+export const object = (options = {}) => ({
+  validator: value => typeof value === "object",
+  options
+});
+
+export const lint = (options = {}) => ({ options });
