--- src/tests/index.js
@@ -1,1 +1,0 @@
-export default (...args) => fun => fun(...args).validator;
