+++ src/utils.js
@@ -1,17 +1,19 @@
 export const getErrorMessage = options =>
   options.message !== undefined ? options.message : false;
 
-export const contextWrapper = rules => context => {
-  const newContext = context;
+export const fieldWrapper = rules => field => {
+  const newField = field;
   Object.entries(rules).forEach(([name, rule]) => {
-    newContext.prototype[name] = function(...args) {
+    newField.prototype[name] = function(...args) {
       return this._add(rule(...args));
     };
   });
-  return newContext;
+  return newField;
 };
 
 export const isObjectsEquals = (obj1, obj2) =>
   JSON.stringify(obj1) === JSON.stringify(obj2);
 
 export const isEmpty = obj => isObjectsEquals(obj, {});
+
+export const getValidator = (...args) => fun => fun(...args).validator;
